criteriaresultentry now only taking a criteria id
